Move `ApiLock` to types.ts
